Day 24: Never Tell Me The Odds

+++ year2015/src/test/kotlin/year2015/Day7Test.kt
@@ -4,12 +4,12 @@ import framework.DayTest
 
 class Day7Test : DayTest(Day7, {
     part1 {
-        example = (72).toUShort()
-        input = (956).toUShort()
+        example = 72
+        input = 956
     }
 
     part2 {
-        example = (72).toUShort()
-        input = (40149).toUShort()
+        example = 72
+        input = 40149
     }
 })
